% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{read_primers}
\alias{read_primers}
\title{Read primer from CSV}
\usage{
read_primers(file = get_golem_config("bq_primer_file"))
}
\arguments{
\item{file}{Character. Path to the CSV file containing primer definitions, with rows
prefixed by `#`.}
}
\value{
A data frame with all primer properties.
}
\description{
Read primer from CSV
}
\examples{
\dontrun{
primers <- read_primer("primers.csv")
print(primers)
}

}
