% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_barque_config}
\alias{write_barque_config}
\title{Write a Barque Configuration File}
\usage{
write_barque_config(
  file = get_golem_config("bq_config_file"),
  NCPUS = get_golem_config("bq_ncpus"),
  PRIMER_FILE = get_golem_config("bq_primer_file"),
  SKIP_DATA_PREP = 0,
  CROP_LENGTH = 200,
  MIN_OVERLAP = 30,
  MAX_OVERLAP = 280,
  MAX_PRIMER_DIFF = 8,
  SKIP_CHIMERA_DETECTION = 0,
  MAX_ACCEPTS = 20,
  MAX_REJECTS = 20,
  QUERY_COV = 0.6,
  MIN_HIT_LENGTH = 100,
  MIN_HITS_SAMPLE = 10,
  MIN_HITS_EXPERIMENT = 20,
  NUM_NON_ANNOTATED_SEQ = 200,
  MIN_DEPTH_MULTI = 10,
  SKIP_OTUS = 1,
  MIN_SIZE_FOR_OTU = 20
)
}
\arguments{
\item{file}{Character. Output file path for the `.sh` config script.}

\item{NCPUS}{Integer. Number of CPUs to use. A lot of the steps are parallelized (int, 1+).}

\item{PRIMER_FILE}{Character. File with PCR primers information.}

\item{SKIP_DATA_PREP}{Integer. Set to 1 to skip data preparation steps, 0 to run the full pipeline (recommended).}

\item{CROP_LENGTH}{Integer. Cut reads to this length after filtering. Should be just under amplicon length.}

\item{MIN_OVERLAP}{Integer. Minimum number of overlapping nucleotides to merge reads (int, 1+).}

\item{MAX_OVERLAP}{Integer. Maximum number of overlapping nucleotides to merge reads (int, 1+).}

\item{MAX_PRIMER_DIFF}{Integer. Maximum number of differences allowed between primer and sequence (int, 0+).}

\item{SKIP_CHIMERA_DETECTION}{Integer. 0 to search for chimeras (recommended), 1 to skip detection or reuse existing files.}

\item{MAX_ACCEPTS}{Integer. Accept at most this number of sequences before stopping search (int, 1+).}

\item{MAX_REJECTS}{Integer. Reject at most this number of sequences before stopping search (int, 1+).}

\item{QUERY_COV}{Numeric. At least that proportion of the sequence must match the database (float, 0-1).}

\item{MIN_HIT_LENGTH}{Integer. Minimum vsearch hit length to keep in results (int, 1+).}

\item{MIN_HITS_SAMPLE}{Integer. Minimum number of hits in at least one sample to keep in results (int, 1+).}

\item{MIN_HITS_EXPERIMENT}{Integer. Minimum number of hits in the whole experiment to keep in results (int, 1+).}

\item{NUM_NON_ANNOTATED_SEQ}{Integer. Number of most frequent non-annotated unique reads to keep (int, 1+).}

\item{MIN_DEPTH_MULTI}{Integer. Minimum depth to report unique reads per sample in multiple-hit reports.}

\item{SKIP_OTUS}{Integer. Set to 1 to skip OTU creation, 0 to use it.}

\item{MIN_SIZE_FOR_OTU}{Integer. Minimum coverage of unique reads used for OTUs.}
}
\value{
Invisibly returns `TRUE` on success. Writes a Bash script to the specified output path.
}
\description{
Generates a `.sh` file to configure a Barque pipeline run, with all parameters inserted as shell variables.
The function ensures reproducibility and simplifies launching Barque with consistent settings.
}
\examples{
\dontrun{
write_barque_config("barque_config.sh", NCPUS = 16, CROP_LENGTH = 180)
}
}
